- Rename module from pymeta3 to pymeta - setup.py now installs correctly - "python -m venv python; python/bin/python setup.py install; python/bin/python tests/test_pymeta.py"
