+++ .gitignore
@@ -1,3 +1,4 @@
 *.swp
 *.pyc
 tags
+build/*
