+++ setup.py
@@ -1,0 +1,12 @@
+#
+
+from distutils.core import setup
+
+setup(
+    name="pymeta",
+    version="1.1",
+    packages=[
+        "pymeta",
+    ],
+)
+
